# app/http/swagger/swagger.yaml
basePath: /
definitions:
  base.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  demo.UserDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  user.ChangeAmountParam:
    properties:
      amount:
        type: integer
      user_id:
        type: integer
    type: object
info:
  contact:
    email: yejianfeng
    name: yejianfeng1
  description: 这个模块负责显示hade的接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/swaggo/swag
  title: hade
  version: "1.1"
paths:
  /demo/demo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/demo.UserDTO'
              type: array
            type: array
      summary: 获取所有用户
      tags:
      - demo
  /demo/demo2:
    get:
      description: 获取所有学生,不进行分页
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/demo.UserDTO'
            type: array
      summary: 获取所有学生
      tags:
      - demo
  /user/amount:
    post:
      consumes:
      - application/json
      description: 更改金额
      operationId: ChangeAmount
      parameters:
      - description: 查询详情请求参数
        in: body
        name: ChangeAmountParam
        required: true
        schema:
          $ref: '#/definitions/user.ChangeAmountParam'
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功的流程定义数据
          schema:
            $ref: '#/definitions/base.Result'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/base.Result'
      summary: 更改金额
      tags:
      - ChangeAmount
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"
x-extension-openapi:
  example: value on a json format